requirements.txt
requests==2.27.1
boto3==1.20.29
botocore==1.23.29
pyyaml==6.0
kafka-python==2.0.2
python-snappy==0.6.0
lz4==3.1.10
tzlocal==3.0
mysql-connector-python==8.0.27
numpy==1.22.3
pandas==1.4.1
